requirement.txt
numpy
pytorch
opencv-python
PIL
tqdm
pandas